Port EscoriasSMP fork features to 26.2
- Add announce mode (MicPacket announce flag, AnnouncementSoundPacket, voicechat announce command, announce key/PTT)
- Make getSender/getSenderUnconnected O(1) via connection index
- Performance: packet queue cap, TalkCache/AudioPacketBuffer/ALSpeakerBase/CooldownTimer optimizations
- Bump version to 2.6.26+26.2
- Add build-and-release workflow (Fabric/Forge/NeoForge/Paper/Quilt)

+++ common/src/main/java/de/maxhenkel/voicechat/voice/common/AudioUtils.java
@@ -73,7 +73,7 @@ public class AudioUtils {
 
         short[] shorts = new short[floats.length];
         for (int i = 0; i < floats.length; i++) {
-            shorts[i] = ((Float) (floats[i] * scale)).shortValue();
+            shorts[i] = (short) (floats[i] * scale);
         }
         return shorts;
     }
@@ -81,7 +81,7 @@ public class AudioUtils {
     public static float[] shortsToFloats(short[] shorts) {
         float[] floats = new float[shorts.length];
         for (int i = 0; i < shorts.length; i++) {
-            floats[i] = ((Short) shorts[i]).floatValue();
+            floats[i] = shorts[i];
         }
         return floats;
     }
@@ -89,7 +89,7 @@ public class AudioUtils {
     public static byte[] floatsToBytes(float[] floats) {
         byte[] bytes = new byte[floats.length * 2];
         for (int i = 0; i < floats.length; i++) {
-            short x = ((Float) floats[i]).shortValue();
+            short x = (short) floats[i];
             bytes[i * 2] = (byte) (x & 0x00FF);
             bytes[i * 2 + 1] = (byte) ((x & 0xFF00) >> 8);
         }
